+++ components/singleactivity/impl/src/main/java/com/flipperdevices/singleactivity/impl/SingleActivity.kt
@@ -28,6 +28,7 @@ import com.flipperdevices.selfupdater.api.SelfUpdaterApi
 import com.flipperdevices.singleactivity.impl.composable.ComposableSingleActivityNavHost
 import com.flipperdevices.singleactivity.impl.di.SingleActivityComponent
 import com.flipperdevices.singleactivity.impl.utils.AppOpenMetricReported
+import com.flipperdevices.unhandledexception.api.UnhandledExceptionApi
 import kotlinx.collections.immutable.toPersistentSet
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -55,11 +56,16 @@ class SingleActivity :
     @Inject
     lateinit var appOpenMetricReported: AppOpenMetricReported
 
+    @Inject
+    lateinit var unhandledExceptionApi: UnhandledExceptionApi
+
     private var globalNavController: NavHostController? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         ComponentHolder.component<SingleActivityComponent>().inject(this)
+
+        unhandledExceptionApi.initExceptionHandler()
 
         info {
             "Create new activity with hashcode: ${this.hashCode()} " +