+++ components/unhandledexception/api/src/main/kotlin/com/flipperdevices/unhandledexception/api/UnhandledExceptionApi.kt
@@ -1,0 +1,9 @@
+package com.flipperdevices.unhandledexception.api
+
+import kotlinx.coroutines.flow.Flow
+
+interface UnhandledExceptionApi {
+    fun initExceptionHandler()
+
+    fun isBleConnectionForbiddenFlow(): Flow<Boolean>
+}